Added my Schools Name
Im a 2 Years Trainee from Germany currently learning Application Development at a School for adult Education. I Hope for getting access to be able, to learn Programming in Java with the professional JetBrains IDE's and Tools. I would get in touch with my Teachers when Access is Granted, so the other Students get to use the JB Tools, too.